% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_values.R
\name{analyze_gust_factor}
\alias{analyze_gust_factor}
\title{Analyze Gust Factor}
\usage{
analyze_gust_factor(data, min_wind_speed = 5)
}
\arguments{
\item{data}{Data frame with gust and wind_speed columns}

\item{min_wind_speed}{Minimum sustained wind speed to consider (default: 5 m/s)}
}
\value{
List with:
\itemize{
\item summary: summary statistics of gust factor
\item extreme_gusts: observations with high gust factors
\item by_wind_category: gust factor by wind speed category
}
}
\description{
Analyzes the ratio of peak gust to sustained wind speed.
This is the wind equivalent of the wave Hmax/Hs ratio.
}
\examples{
\dontrun{
con <- connect_duckdb()
data <- query_buoy_data(con, variables = c("time", "wind_speed", "gust"))
gust_analysis <- analyze_gust_factor(data)
DBI::dbDisconnect(con)
}
}
