% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_parquet.R
\name{analyze_parquet_storage}
\alias{analyze_parquet_storage}
\title{Analyze Parquet Storage}
\usage{
analyze_parquet_storage(data_path = "inst/extdata/parquet")
}
\arguments{
\item{data_path}{Base path for Parquet files}
}
\description{
Get statistics about Parquet file storage.
}
