% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_alert.R
\name{beaufort_to_description}
\alias{beaufort_to_description}
\title{Convert Beaufort Number to Description}
\usage{
beaufort_to_description(beaufort)
}
\arguments{
\item{beaufort}{Integer vector of Beaufort numbers (0-12).}
}
\value{
Character vector of Beaufort descriptions.
}
\description{
Maps Beaufort scale integers (0-12) to standard descriptions.
}
\examples{
beaufort_to_description(0:12)
}
\seealso{
Other storm-alert: 
\code{\link{create_storm_alert_email}()},
\code{\link{detect_storm_events}()},
\code{\link{fetch_all_forecasts}()},
\code{\link{fetch_all_marine_forecasts}()},
\code{\link{fetch_met_eireann_warnings}()},
\code{\link{fetch_open_meteo_forecast}()},
\code{\link{fetch_open_meteo_marine}()},
\code{\link{knots_to_beaufort}()},
\code{\link{p_hmax_exceedance}()},
\code{\link{send_storm_alert}()},
\code{\link{summarise_forecast_rogue_risk}()}
}
\concept{storm-alert}
