% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_science.R
\name{calculate_hs_from_elevation}
\alias{calculate_hs_from_elevation}
\title{Calculate Significant Wave Height from Raw Elevations}
\usage{
calculate_hs_from_elevation(elevations)
}
\arguments{
\item{elevations}{Numeric vector of surface elevation measurements (m)}
}
\value{
Significant wave height in meters
}
\description{
Calculates Hs from a time series of surface elevation measurements
using the spectral method (4 * sigma).
}
\examples{
# Simulated wave elevation time series
t <- seq(0, 1000, by = 0.5)  # 1000 seconds at 2Hz
elevation <- 0.5 * sin(2*pi*t/8) + 0.3 * sin(2*pi*t/12) + rnorm(length(t), 0, 0.1)
hs <- calculate_hs_from_elevation(elevation)
}
