% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_values.R
\name{compare_rogue_wave_gust}
\alias{compare_rogue_wave_gust}
\title{Compare Rogue Wave and Rogue Gust Occurrence}
\usage{
compare_rogue_wave_gust(data)
}
\arguments{
\item{data}{Data frame with wave_height, hmax, wind_speed, gust columns}
}
\value{
Data frame comparing occurrence rates
}
\description{
Compares the occurrence rates of rogue waves (Hmax/Hs > 2) and
extreme gusts (gust/wind > 2.6).
}
\examples{
data <- data.frame(
  wave_height = c(3, 4, 5, 2.5),
  hmax = c(5, 9, 8, 4),
  wind_speed = c(10, 15, 20, 8),
  gust = c(15, 40, 30, 12)
)
compare_rogue_wave_gust(data)
}
