% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connect_duckdb}
\alias{connect_duckdb}
\title{Create or Connect to Irish Buoys DuckDB Database}
\usage{
connect_duckdb(db_path = "inst/extdata/irish_buoys.duckdb", create_new = FALSE)
}
\arguments{
\item{db_path}{Character, path to database file (default: "inst/extdata/irish_buoys.duckdb")}

\item{create_new}{Logical, whether to create new database (default: FALSE)}
}
\value{
DBI connection object to the DuckDB database
}
\description{
Creates a new DuckDB database or connects to an existing one for storing
Irish Weather Buoy Network data. Sets up the schema if creating new.
}
\examples{
# Connect to existing database
con <- connect_duckdb()

# Create new database
con <- connect_duckdb(create_new = TRUE)

# Don't forget to disconnect when done
DBI::dbDisconnect(con)
}
