% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_annual_trends}
\alias{create_plot_annual_trends}
\title{Create Annual Trends Line Plot}
\usage{
create_plot_annual_trends(annual_trends, date_caption = NULL)
}
\arguments{
\item{annual_trends}{Annual trends from calculate_annual_trends}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Annual Trends Line Plot
}
