% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_gust_by_category}
\alias{create_plot_gust_by_category}
\title{Create Gust Factor by Category Plot}
\usage{
create_plot_gust_by_category(gust_analysis, date_caption = NULL)
}
\arguments{
\item{gust_analysis}{Gust factor analysis results}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Gust Factor by Category Plot
}
\examples{
\dontrun{
gust <- list(
  by_station_category = data.frame(
    station_id = rep(c("M2", "M3"), each = 3),
    wind_category = rep(c("0-10", "10-20", "20+"), 2),
    mean_gf = runif(6, 1.1, 1.8),
    p95_gf = runif(6, 1.5, 2.5),
    n = sample(50:500, 6)
  )
)
create_plot_gust_by_category(gust)
}
}
