% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_monthly_wind}
\alias{create_plot_monthly_wind}
\title{Create Monthly Wind Speed Bar Plot}
\usage{
create_plot_monthly_wind(seasonal_means_wind, date_caption = NULL)
}
\arguments{
\item{seasonal_means_wind}{Seasonal means from calculate_seasonal_means}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Monthly Wind Speed Bar Plot
}
