% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_return_levels}
\alias{create_plot_return_levels}
\title{Create Return Levels Plot}
\usage{
create_plot_return_levels(
  return_levels,
  variable = "wave",
  date_caption = NULL
)
}
\arguments{
\item{return_levels}{Return levels data frame}

\item{variable}{Variable name for title ("wave", "wind", or "hmax")}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Return Levels Plot
}
