% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_rogue_all}
\alias{create_plot_rogue_all}
\title{Create Rogue Wave All Stations Plot}
\usage{
create_plot_rogue_all(rogue_events)
}
\arguments{
\item{rogue_events}{Data frame of rogue wave events}
}
\value{
plotly object
}
\description{
Create Rogue Wave All Stations Plot
}
