% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_rogue_by_station}
\alias{create_plot_rogue_by_station}
\title{Create Rogue Wave By Station Subplot}
\usage{
create_plot_rogue_by_station(rogue_events, date_caption = NULL)
}
\arguments{
\item{rogue_events}{Data frame of rogue wave events}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Rogue Wave By Station Subplot
}
