% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_rogue_gusts_all}
\alias{create_plot_rogue_gusts_all}
\title{Create Rogue Gusts All Stations Plot}
\usage{
create_plot_rogue_gusts_all(rogue_gust_events)
}
\arguments{
\item{rogue_gust_events}{Data frame of rogue gust events}
}
\value{
plotly object
}
\description{
Create Rogue Gusts All Stations Plot
}
\examples{
\dontrun{
gusts <- data.frame(
  time = as.POSIXct("2024-01-01") + (1:10) * 3600,
  station_id = rep(c("M2", "M3"), 5),
  gust_ratio = runif(10, 1.5, 3.0),
  gust = runif(10, 20, 50),
  wind_speed = runif(10, 10, 25),
  wave_height = runif(10, 2, 6)
)
create_plot_rogue_gusts_all(gusts)
}
}
