% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_time_of_day}
\alias{create_plot_time_of_day}
\title{Create Time of Day Bar Plot}
\usage{
create_plot_time_of_day(rogue_conditions, date_caption = NULL)
}
\arguments{
\item{rogue_conditions}{Data frame with rogue wave conditions}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Time of Day Bar Plot
}
