% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_week_of_year}
\alias{create_plot_week_of_year}
\title{Create Week of Year Stacked Bar Plot}
\usage{
create_plot_week_of_year(rogue_conditions, date_caption = NULL)
}
\arguments{
\item{rogue_conditions}{Data frame with rogue wave conditions}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Week of Year Stacked Bar Plot
}
