% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_wind_beaufort}
\alias{create_plot_wind_beaufort}
\title{Create Wind Speed by Beaufort Scale Plot}
\usage{
create_plot_wind_beaufort(rogue_conditions, date_caption = NULL)
}
\arguments{
\item{rogue_conditions}{Data frame with rogue wave conditions}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Wind Speed by Beaufort Scale Plot
}
\examples{
\dontrun{
conditions <- data.frame(
  time = as.POSIXct("2024-01-01") + (1:10) * 3600,
  station_id = rep(c("M2", "M3"), 5),
  wind_speed = runif(10, 5, 35),
  rogue_ratio = runif(10, 2.0, 2.5),
  hmax = runif(10, 8, 15), wave_height = runif(10, 3, 6)
)
create_plot_wind_beaufort(conditions)
}
}
