% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_analysis.R
\name{haversine_distance}
\alias{haversine_distance}
\title{Calculate Distance Between Two Stations}
\usage{
haversine_distance(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1, lon1}{Coordinates of first point (degrees)}

\item{lat2, lon2}{Coordinates of second point (degrees)}
}
\value{
Distance in kilometers
}
\description{
Calculates the great-circle distance between two points using the Haversine formula.
}
\examples{
# Distance from M6 to M2
haversine_distance(53.07, -15.93, 51.22, -9.99)
}
