% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_alert.R
\name{knots_to_beaufort}
\alias{knots_to_beaufort}
\title{Convert Wind Speed in Knots to Beaufort Scale}
\usage{
knots_to_beaufort(wind_speed_kn)
}
\arguments{
\item{wind_speed_kn}{Numeric vector of wind speeds in knots.}
}
\value{
Integer vector of Beaufort numbers (0-12).
}
\description{
Vectorized conversion from wind speed in knots to the Beaufort scale (0-12).
}
\examples{
knots_to_beaufort(c(0, 5, 20, 34, 48, 64))
}
\seealso{
Other storm-alert: 
\code{\link{beaufort_to_description}()},
\code{\link{create_storm_alert_email}()},
\code{\link{detect_storm_events}()},
\code{\link{fetch_all_forecasts}()},
\code{\link{fetch_met_eireann_warnings}()},
\code{\link{fetch_open_meteo_forecast}()},
\code{\link{send_storm_alert}()}
}
\concept{storm-alert}
