% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{log_update}
\alias{log_update}
\title{Log Data Update}
\usage{
log_update(con, start_date, end_date, records_added, stations, notes = NULL)
}
\arguments{
\item{con}{DBI connection object}

\item{start_date}{Start date of updated data}

\item{end_date}{End date of updated data}

\item{records_added}{Number of records added}

\item{stations}{Comma-separated list of updated stations}

\item{notes}{Optional notes about the update}
}
\value{
Invisible NULL
}
\description{
Records information about data updates to the update_log table.
}
\keyword{internal}
