% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_analysis.R
\name{station_distance_matrix}
\alias{station_distance_matrix}
\title{Calculate Distance Matrix Between All Stations}
\usage{
station_distance_matrix(station_info = NULL)
}
\arguments{
\item{station_info}{Data frame from get_station_info() or NULL to use default}
}
\value{
Named matrix of distances in km
}
\description{
Creates a matrix of distances between all station pairs.
}
\examples{
station_distance_matrix()
}
