% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_model.R
\name{wave_model}
\alias{wave_model}
\title{Wave Height Prediction Model}
\description{
Functions for building and using a Random Forest model to predict
significant wave height from meteorological variables.
}
