% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_plumber.R
\name{api_plumber}
\alias{api_plumber}
\title{Plumber API for irishbuoys}
\description{
Functions for running a live REST API that serves pre-built
static JSON data from the targets pipeline.
}
\seealso{
Other api: 
\code{\link{api_static}},
\code{\link{create_api_router}()},
\code{\link{generate_api_decomposition}()},
\code{\link{generate_api_extremes}()},
\code{\link{generate_api_gust_factors}()},
\code{\link{generate_api_index}()},
\code{\link{generate_api_latest}()},
\code{\link{generate_api_methods}()},
\code{\link{generate_api_sources}()},
\code{\link{generate_api_spatial}()},
\code{\link{generate_api_status}()},
\code{\link{generate_api_trends}()},
\code{\link{run_api}()}
}
\concept{api}
