% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_gust_by_category}
\alias{create_plot_gust_by_category}
\title{Create Gust Factor by Category Plot}
\usage{
create_plot_gust_by_category(gust_analysis, date_caption = NULL)
}
\arguments{
\item{gust_analysis}{Gust factor analysis results}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Gust Factor by Category Plot
}
