% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_gusts_vs_waves}
\alias{create_plot_gusts_vs_waves}
\title{Create Rogue Gusts vs Rogue Waves Scatter Plot}
\usage{
create_plot_gusts_vs_waves(analysis_data)
}
\arguments{
\item{analysis_data}{Full analysis data with both ratios computed}
}
\value{
plotly object
}
\description{
Create Rogue Gusts vs Rogue Waves Scatter Plot
}
