% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_monthly_wave}
\alias{create_plot_monthly_wave}
\title{Create Monthly Wave Height Bar Plot}
\usage{
create_plot_monthly_wave(seasonal_means_wave, date_caption = NULL)
}
\arguments{
\item{seasonal_means_wave}{Seasonal means from calculate_seasonal_means}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Monthly Wave Height Bar Plot
}
