% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_rogue_by_station}
\alias{create_plot_rogue_by_station}
\title{Create Rogue Wave By Station Subplot}
\usage{
create_plot_rogue_by_station(rogue_events, date_caption = NULL)
}
\arguments{
\item{rogue_events}{Data frame of rogue wave events}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Rogue Wave By Station Subplot
}
\examples{
\dontrun{
rogue <- data.frame(
  time = as.POSIXct("2024-01-01") + (1:10) * 3600,
  station_id = rep(c("M2", "M3"), 5),
  rogue_ratio = runif(10, 2.0, 2.5),
  hmax = runif(10, 8, 15), wave_height = runif(10, 3, 6),
  wind_speed = runif(10, 10, 30)
)
create_plot_rogue_by_station(rogue)
}
}
