% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_rogue_gusts}
\alias{create_plot_rogue_gusts}
\title{Create Rogue Gusts by Station Plot}
\usage{
create_plot_rogue_gusts(gust_analysis)
}
\arguments{
\item{gust_analysis}{Gust factor analysis results}
}
\value{
plotly object
}
\description{
Create Rogue Gusts by Station Plot
}
