% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_rogue_gusts_all}
\alias{create_plot_rogue_gusts_all}
\title{Create Rogue Gusts All Stations Plot}
\usage{
create_plot_rogue_gusts_all(rogue_gust_events)
}
\arguments{
\item{rogue_gust_events}{Data frame of rogue gust events}
}
\value{
plotly object
}
\description{
Create Rogue Gusts All Stations Plot
}
