% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_stl}
\alias{create_plot_stl}
\title{Create STL Decomposition Plot}
\usage{
create_plot_stl(wave_stl, date_caption = NULL)
}
\arguments{
\item{wave_stl}{STL decomposition from calculate_wave_seasonality}

\item{date_caption}{Date range caption}
}
\value{
ggplot2 object
}
\description{
Create STL Decomposition Plot
}
