% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_time_of_day}
\alias{create_plot_time_of_day}
\title{Create Time of Day Bar Plot}
\usage{
create_plot_time_of_day(rogue_conditions, date_caption = NULL)
}
\arguments{
\item{rogue_conditions}{Data frame with rogue wave conditions}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Time of Day Bar Plot
}
\examples{
\dontrun{
conditions <- data.frame(
  time = as.POSIXct("2024-01-01") + (1:20) * 3600,
  time_of_day = rep(c("Morning", "Afternoon", "Evening", "Night"), 5),
  rogue_ratio = runif(20, 2.0, 2.5)
)
create_plot_time_of_day(conditions)
}
}
