% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{create_plot_wind_beaufort}
\alias{create_plot_wind_beaufort}
\title{Create Wind Speed by Beaufort Scale Plot}
\usage{
create_plot_wind_beaufort(rogue_conditions, date_caption = NULL)
}
\arguments{
\item{rogue_conditions}{Data frame with rogue wave conditions}

\item{date_caption}{Date range caption}
}
\value{
plotly object
}
\description{
Create Wind Speed by Beaufort Scale Plot
}
